ADD http request library, API url config
Vue documentation recommends the axios library for handling http
requests, so that's what we're using. I've attached axios to the Vue
object so that it should be available globally in all our Vue
components.

Now that we need to make requests to the API, we need to configure the
API url. This is done using the environment variable file
`.env.development`. It's recommended not to check in the environment
files, but we need it for docker-compose to work as we mount the entire
ui/ directory into the container. Developers who aren't using docker
will need to modify it.

Also changed docker-compose so that the ui will serve on the same port
that docker-compose exposes, to reduce confusion with the ui log output.