ADD reverse proxy to avoid CORS issues
Since the UI and API are running separately, on different ports in
docker-compose, browsers will error out on API requests cause it thinks
a cross-origin attack is in progress. To avoid this, added the reverse
proxy Traefik to docker-compose. It'll redirect requests to the UI on
the /api path to the API.

However, this breaks the autogenerated documentation provided by FastAPI
since those documentation assumes they're being served at the root path.
So direct access to the API is still available.