ADD starting basic user creation, API url config
Trying to abstract user edit and user add forms into one user info form.
Also put together a basic user add page with an API call so we can start
working on the API side of things.

The API url is configured using Vue cli's environment variables. This is
the `.env.development` file which currently just defines the URL as
being located as the /api/ endpoint. This, by default, works with the
docker-compose setup but will need to be customized on manual setups.