ADD initial db support using sqlalchemy & alembic
SQLAlchemy is used to talk to the postgres database.

Alembic used for migrating the database schema between versions.

The docker startup script `prestart.sh` will now  run alembic migrations
to bring the database up to spec.

While the alembic documentation finds creating the database from scratch
by running all migrations to be overkill, we don't really have much of a
database yet, so I'm not going to start worrying about the time cost of
running all the migrations yet.

The downside of alembic migrations seems to be keeping them in sync with
the models. There's essentially two independent database schema specs,
one that exists in the models and one in the migrations. We'll need to
manually ensure they're in agreement with each other.

As environmental variables are the easiest way for us to pass
configuration in kubernetes, I've setup the database configuration to
take their values from environmental variables: DB_USER, DB_PASSWORD,
DB_HOST, DB_NAME

Currently, there is only 1 table, the users table.