ADD api async db operations, list users & add user
Instead of the using SQLAlchemy's ORM, now using the more low level EL.
This is to benefit from async operation. The terribly named 'databases'
python package enables async on SQLAlchemy's EL base but has no plans on
enabling async on SQLAlchemy ORM due to complexity.

EL does require us to do more manual management, but it should be faster
when executing than the ORM.

Sending a GET request to /users/ will return a list of users.

Sending a POST request to /users/ will create a new user.

The request and response data is validated against data models to ensure
we get the data we need and that we don't send data that shouldn't be
sent.