ADD user listing, vue-mc, tailwind css framework
Vue-mc is for data models on the UI side. It encapsulates axios requests
for us, so we don't have to handle manual model saving/validation.

The UserTable component lists users in the database. I'm not sure if
UserTable or the other UserInfoForm components should be in components
or views yet. They are listed in router.js and might be better organized
as a view.

Tailwind is a new to me approach to CSS. Where they don't provide
default styles but give you a lot of utility classes that you can
combine to make your own. Not sure if it'll work out, but trying it.