ADD layout system to the UI
In preparation for the login page where users should only see the login
form and not the sidebar, needed to have a way of displaying different
layouts on different pages. Found a nice tutorial at:
https://itnext.io/anyway-heres-how-to-create-a-multiple-layout-system-with-vue-and-vue-router-b379baa91a05

You basically specify what layout to use in the router's path's meta
field. This is implemented for the About page as a demonstration. The
layout components are made globally available in main.js and will be
loaded by App.vue depending on the layout name. If no layout is
provided, it defaults to the layout with the sidebar.

I wanted the layout you give to the router to be same as the layout
component's filename. So the DefaultLayout is just "Default" and the
NoSidebarLayout is just "NoSidebar".