ADD login form, password field component
Since the password field has additional functionality and is used in
user info and login, it needs to be placed into a component to reduce
code duplication. This was interesting, as we needed to use v-model
binding on the password field, so the component version needed a bit of
extra work to keep that working.

One major issue encountered trying to make the show/hide password
functionality work is that Vue was destroying the input and recreating
it, so we lost the previously entered password everytime we toggle the
show/hide. The work around for this seems to be storing a duplicate of
the password in data.

Also trying to make a standardized vertical form, not sure where that'll
go.