ADD LabelledIcon utility component
LabelledIcon takes in an icon component and a label prop. The label only
appears on md or larger screens. So on smaller mobile devices, they'll
only see a slightly larger icon.

Also realized why my global.css import was messing everything up unless
it was at the end of the file. Turns out postcss still requires
semicolons at the end of every line. I got out of the habit of doing
that even in JS, so didn't realize until it was messing up with the
material design icon's css import.