ADD API initial token generation on sign in
Give the user a token when they sign in. Still working on verifying the
token on return visits.

The token generation is done using pyjwt. Three new environment variable
configurations are now available, although only the secret is required:

TOKEN_SECRET - Required, used to sign the token, can be generated with
'openssl rand -hex 32'. If not provided, will warn via error logs but
will use an insecure pre-generated string.

TOKEN_ALGORITHM - Optional. Defaults to "HS256".

TOKEN_EXPIRE_MINUTES - Optional. Number of minutes before the token
expires, defaults to 180.

Reading up on a few JWT resources, it looks like I've been getting it
wrong. The token shouldn't be used as a session,  it shouldn't be used
to store the type of session data that sessions are known for. So I've
renamed Session to Token. Since we're only treating it as a user
identifier, not as a way to store user session data.

Added a migration to make usernames unique. Realized a bit late that
usernames had to be unique when I created two users with the same
username while testing sign in.