ADD UI sign in, sign out, token storage
Users can now sign in and sign out. The token sent by the API on sign in
is now properly stored in vuex.

Vuex has a module system, so I've created the store/ directory to put
Vuex modules in. Currently only the auth module is in there.

There is now an error component for displaying errors.