EDIT UI seamless sign in and sign out
Got rid of the dedicated sign in page. Sign in is now just a component
included in the whole app which will be visible if the user needs to be
logged in. This should prevent data loss for users whose tokens expire
in the middle of doing stuff, as their stuff is just temporarily hidden
in favour of the sign in form.

There is now two home pages. One for not signed in users, which is
served at the web root. And one for signed in users, served at /home

There is now a sign out confirmation page. Going by general UI
recommendation to explicitly confirm to the user that sign out is
successful.