EDIT UI sign out, home page behaviour if signed in
When signed in, users visiting the home page should be directed to the
signed in home page.

When signed in, users who somehow ended up on the signed out
acknowledgement page should be asked if they really wanted to sign out
or if they just ended up there by accident.