REMOVE UI sidebar layout, using top nav for now
The original idea was to have a conventional side nav that's toggled on
and off with a hamburger menu on mobile. But reading up on it, there's
some interesting arguments against using hamburger menus. So going to
try something different.

Current plan is to replace them with tabs or pill based nav and see how
far we can go.

Also moved the user pages to the admin section, as that's where they're
supposed to go anyway.