EDIT UI admin/user lists users in card format
Previously listing users using a table, which seems impossible to get
working right on mobile, so now trying to list them in cards and see if
that'll work out better.

Also added proper error handling and display to the UserTable component.

Realized that we're not properly handling expired tokens yet, so making
this commit and then going to work on that.