EDIT UI sign user out on token expiry
If a request gets rejected with a 401, the user's token probably expired
and we need to get the user to sign in again.

This is done with an axios interceptor that check response for 401
errors. Strangely, I couldn't get the interceptor to stick on the
instance variable. It'll only work if I put the interceptor on the
global axios, which the instance variable then inherits. Not sure why,
probably arcacne JS stuff that I don't have time to research, sadly.