EDIT UI user table improve style on mobile
Was trying to get the user cards on mobile to expand in width to fill
the screen. Turns out I was using flexbox wrong and the layout had extra
divs that was locking the width of the user table. Removed the extra
divs and stuff worked. Looks like flex-1 and flex-auto will both expand
to fill space.

Each user entry will now be displayed with a flex user card on mobile
and css table on desktops. Probably needs more tweaking but probably
better to spend time on other stuff.