ADD API list/create/edit courses
Need to start working on the permission system, but before that, need
something to permission against. Courses are where most of our
permissions are going to be located, so need to have courses and course
enrolment first before starting work on permissions.

There's a lot of very similar code with users and courses. If the code
similarity continues in the future, some refactoring should be done to
reduce the duplicate code.