ADD API user & course delete
Can now delete users and courses. The delete request returns an empty
body and a 204 NO CONTENT since that seems to be the most common
pattern.

Changed the user/course add response code to 201 CREATED, as that seems
to be the standard.

Also added the 'description' and 'isActive' columns to courses after
talking to instructors. The 'isActive' should be helpful in hiding old
courses that are no longer being run, similar to how iPeer does it.