EDIT API merge migrations, rename to singular form
Now using the singular instead of the plural for table names, e.g.:
"user" instead of "users". Just for easier consistency, as I've been
using singular in the python code.

Took the opportunity to also merge the smaller migrations, we don't
really need that many migrations.

Also will create a default user with the username "admin" and password
"admin".