ADD API initial enrolment functions
Enrolment is adding users to courses. This initial commit creates the
enrolment table and associated controllers/models.

There is now also a Role table for querying course roles.

Currently, only functional aspect is getting a list of users in a
course. This involved a somewhat complicated sqlalchemy join query. I
had a lot of issues with the query which turned out to be caused by the
fact that 'user' is a keyword in postgres that gets converted to the
current_user function. So when I tried doing raw queries against what I
think was the user table, it keeps calling current_user. The solution
was to double quote "user" so it gets treated as the table it's supposed
to be. SQLAlchemy does this for you thanksfully.