EDIT UI course add/edit/delete now working
Turned out to be easier than I thought to get add/edit/delete working
for courses, so also went on an minor improvements spree.

Courses returned by the API will now be sorted by name, users by
username.

Added description to Course models on both API and UI side so
description can be properly edited.

Fixed "null" string showing up when a null value is passed from the API.
This seems to be caused by specifying a mutator on the fields in
question, I guess it performs some kind of string conversion that
converts the javascript null into the string "null".

Courses subpages are proper child routes under /admin/courses, will
probably need to do this for the user pages as well.

Added an Info status box for informational messages to the user.

Changed to fractional widths on the default layout to get rid of the
issue of navigational components jumping around on the user due to the
auto widths changing on every page.