ADD UI breadcrumb system
Couldn't find a breadcrumb library that was fast and easy to use.
Primarily because of a lack of documentation on those library's part.
Turns out it was easier to just write my own breadcrumb system using
vue-router.

Each route will need a meta.breadcrumb variable for the label used by
the breadcrumb system. If meta.breadcrumb isn't provided, then that
route is skipped.

Each route's label is currently static. I'd like to have a way to change
the route name on the fly. E.g.: For a course, change it to a course
name when the course data loads. When there's time.

I've started splitting the router into different files to try to make
the router more readable once we get more routes. The new router/
directory in plugins/ are where the subroutes should go. The first one
is the admin routes in plugins/router/admin.js

For organizational reasons, I've moved vuex's store/ dir into plugins/
This hopefully makes more sense since store.js is also in plugins/
I didn't want to have both store/ and router/ in the top level src/