ADD UI course home page
This is the page that users end up on when they click on a course listed
in the signed in home page.

Basic starting point with some mockup nav elements. This is where
instructors can go to the course users page and add/remove users from a
course.

The enrolment page for adding and removing users from a course is still
just a mockup.