FIX UI session expired error on initial sign in
Also fixed not having a token expired error.

Turns out the session expired error shows up even on initial sign in,
since getCourses gets fired when the page loads. I've delayed getting
courses until the user is signed in.

When a token expires and the user needs to be relogin, no error message
is shown, user is just directed to the sign in. Added a special
isTokenRejected variable to the auth store to indicate whether the token
was expired. The sign in form will show an error message if this
variable is true.