ADD API signed in user's context (courses & roles)
I've changed the token again, this time removing storage of username
altogether. The more I think about it, because the username can change,
we really shouldn't use it as unique identifier. So we're now completely
relying on the user id.

A new model, the TokenContext, has the signed in user, the courses
they're enroled in, and the roles they have in those courses. The roles
are mapped by course id to the role's name.

Token.getContext is intended to supplement Token.getCurrentUser as a new
signed in dependency. getContext actually depends on getCurrentUser
itself, so you only need one of them. On each request, we'll make sure
the user is signed in and retrieve the signed in user's context
automatically. For requests that don't need the additional data provided
by the context, they should continue using getCurrentUser (less queries
to the database).

EnrolmentTable's getEnroledCourses was modified so that it returns the
user's role along with the courses.

Since REST is supposed to be stateless, we have to retrieve this
information on every request that depends on it. I'm somewhat worried
aobut doing too many queries in one request, but I'm not sure what we
can do to avoid this.