EDIT API courses operations now check permissions
Only relying on Permissions.acl for permissions inside enroled courses.
This is done with the Courses.can function. Some permissions, such as
creating courses, do not use the can() check and instead only check if
the user is an admin. Others, such as getAll(), allows everyone to use,
but returns different results depending on your permission level.

Changed ACL to disallow instructors from deleting courses.

Just realized that id could be an optional parameter in the models
because it's given in the path. Some quick experimentation shows that
CourseTable depends on the id in the model, so making it optional might
require more work than expected.