EDIT API models don't require id given in url
If the API url is /courses/{courseId}, then giving the id in the payload
should be optional, as we could just use the id in the URL. This concept
is a bit harder to map to enrolment since there's several ids acting as
foreign keys, so I've put off dealing with enrolments to think about it
more.

The new ModelChecker helper helps ensures that missing ids are filled in
and that we don't have a mismatch.

Working with the enrolment API, I realized that it differs form the
other APIs in that it expects to operate on lists of enrolments.  This
is a bit jarring, so probably need to overhaul that too.