ADD API get list of roles
The roles only apply inside a course. The only system wide role
available right now is system admin and that is not included in this
list.

It shouldn't be necessary to restrict access to roles, so no permissions
checking is done. It helps that the roles are read only, no
add/edit/deletes.