ADD UI notification & global error system.
Notifications disappear after a short time and thus is good for
success/info messages. Errors might contain information that needs to be
referenced and so shouldn't disappear. So there's now a GlobalError
component that combined with a error vuex store that handles display of
errors. New errors added to the error store is automatically displayed.
A short error notification will appear and a longer, more detailed error
message is displayed at the top of the page.

Errors are reset when the user navigates away from the page. Errors can
also be dismissed manually by the user.

The centralized error handling helps to keep error handling code from
duplicated in every component. Now we can just modify the error vuex
store to handle new response codes and such.

Note that SignIn is NOT using the vuex error or the GlobalError
component. This is because GlobalError was placed in the Default
template, which is hidden when the SignIn is shown.