EDIT API reduce queries when checking items exists
We need to validate ids used in the API path, such as course, exams,
etc.  Instead of doing multiple queries to check if each individual
item exists, this function attempts to shorten it to just 1 query to
check if everything we need exists.

The end goal is to build an SQL query that looks something like this:

  SELECT EXISTS( SELECT 1 FROM course WHERE id=1 ) AND EXISTS( SELECT
FROM exam WHERE id=2 AND course_id=1 ) as "exists";

which returns true if everything exists, false otherwise.

An individual RowSpec would be specified like this:

  RowSpec(ExamTable.table, [ExamTable.table.c.id==2,
                            ExamTable.table.c.course_id==1])

where we specify the table and the fields/values that we need to check.
Note the multiple conditions, we sometimes might want to make sure that
the given exam actually belongs to the course given in the path.