ADD API ExamSource get & upload files
Exam sources are the pdf files where we'll take pages from for mixing
and matching later.

Using filedepot library for easy file management. The problem with
filedepot is that it lacks a way to impose a storage structure. I'd like
files to have a directory structure like course/exam/type. But to impose
this structure would require rolling a custom solution which would take
time. So for now, to get a functional prototype working quickly, we'll
use the library.

Currently just storing files in api/uploaded_files/