EDIT API ExamSource now returns course_id
The UI vuemc models need the course_id to construct the API urls. It was
easier to just do a join query on the API side and bring in the course_id.

Might need to abstract the join query into a generalized version in
TableOp later if we're using it a lot.

Rearranged delete so that it'll delete database entry first before
trying to delete the file. Previously, if the database delete failed,
we'd have lost the file since it was deleted first.