ADD UI working ExamSource upload
Can now upload and delete ExamSource. Later on will need to add
protection and validation:
 * Need to check that page count is same for all sources, warn if not
 * Need to protect ExamSources that've been used from deletion

ExamSourceTable is in it's own component as the ExamForm was getting a
bit long. Need to work out a good way to use vuex to share data. Cause
each component getting its own data is not very efficient.