EDIT API reduce queries used in ExamSources.can
Since the "can" functions are executed for every controller action, we
should try to keep it fast. Instead of doing many queries to validate
that the parent Exam and Course exists, we can reduce it to just 1 query
to validate the ExamSource, Exam, and Course all exists.

This does make the error message less specific, since any one of them
missing would trigger this error.