ADD API ExamComponent basic functions
Can now get/add/edit/delete ExamComponents.

ExamComponents specify how to generate an exam. Each component is an
element on a generated exam paper, such as an ID page, a question, etc.

The API list is getting a bit unwieldy, so I've started using some of
the autogenerated doc features in FastAPI. This should make the API test
page more informative.

Thinking about deprecating the ModelChecker helper. It's faster to just
silently replace the IDs with the ones given in the URL and I'm not sure
what advantage there is in enforcing the IDs to match.