ADD API create/delete exam component sources api
Just a quick implementation of the API for exam component sources. I'm
not sure if this should be it's own separate module or maybe it should be
part of exam components yet. Going to have to see how the UI side goes
before deciding.

There is no edit this time, as I think the workflow on the UI side is
easier to just delete the component source entry and create a new one.