ADD UI see if vuex-orm can replace vuemc
The only issue with vuemc is that it doesn't play well with vuex. So
I'm trying to see if vuex-orm can replace vuemc. I really don't want to
have to recreate the api requests that you get for free with vuemc.

Unfortunately, it's not looking like vuex-orm can properly replace
vuemc. The issue isn't with vuex-orm core, but with the vuex-orm axios
plugin. The plugin doesn't seem to be well maintained and has issues
such as not passing on http errors and not being able to handle errors
in the individual model level. You can specify general error handlers
as a plugin config, but not at the model level.

Looks like we'll just have to go with the workaround where we serialize
vuemc model data to JSON and store that in vuex if we need to share
data.

Makign this commit in case we want to go with vuex-orm later.