EDIT API split requirements.in into base and test
Starting to try to run some automated tests. There are libraries that
are only needed for running automated tests, such as pytest, that
shouldn't be installed in production. To facilitate this split, we need
to have separate requirement files for our dependencies.

It looks like putting the requirements into the requirements/ directory
is the common way to do this. The base requirement file contains only
those needed to run the app. While the test requirement file pulls in
the base and adds libraries that are needed to run test cases.