UPGRADE eslint, yarn keep changing eslint-util ver
Tried to pin eslint-util to >1.4.1 because lower version numbers
triggers a github security alert. But this keeps getting changed by
'yarn install' that downgrades the pinning. So I've just upgraded eslint
since that doesn't seem to break anything. Alternatively, can give the
frozen-lockfile option to 'yarn install' to make it stop tampering with
the lock file.