ADD API basic working test for sign in
First try at a working test case. Testing out the sign in controller.

We're using a real postgres database to run these tests. The postgres
instance is a docker container. It is started by pytest using a pytest
plugin that starts containers according to tests/docker-compose.yml.
This is to avoid mismatch issues with production, as using sqlite or
something similar that doesn't have the full postgres features has
caused some issues in the past.

As starting the docker container on every function or every module can
be pretty expensive, I'm only starting them on every session. The
database is reset instead on every module run by dropping all the tables
and then recreating them. Hopefully, this is faster than recreating the
docker container.

Added basic instructions on how to run the test cases.

Had to move main.py to app/ to make importing it in the test cases easier.