EDIT API split docker into prod & dev versions
The development docker image features live reload & python module
compilation tools that we don't need in the production image.

The production image tries to minimize image size by doing a multistage
build. The first stage compiles the python modules, the compiled modules
are then copied into the final final image. This requires an adjustment
to the pythonpath for gunicorn to run the server on, which is what the
gunicorn_conf.py file is for.

There's also different prestarts for the dev and prod images.

The requirements files now has a development version. This just lets us
skip bringing in pip-tools to production right now.