EDIT UI split dockerfile into dev & prod versions
Part of preparation for deployment. Prod Dockerfile will generate an
image with everything compiled down into a small app that can be served
by nginx.

Also forced eslint errors to kill the prod build if there's an error.
Not sure if this will affect the dev server, I hope not. If it does,
then we'll have to do something else. Interestingly, running the linter
alone doesn't seem trigger the same checks as running the build. I'm not
sure why this is. This is why I'm having the build abort on lint errors,
instead of adding a separate lint step to the gitlab build.

Also changed the environment variable file to be applied to both dev &
prod.  My original plan was to change it based on the kube chart, but we
didn't really end up using the env var file that much.