EDIT UI remove purgecss & postcss-preset-env
PurgeCSS removes unused CSS for you. However, it seems to have caught
the material design icon CSS in a friendly fire. Probably because the
CSS classes doesn't appear in my code, but in code that's pulled in from
npm.

Postcss-preset-env is supposed to convert modern CSS for backward
compatibility with older browsers. I think it was added as part of
PurgeCSS' pipeline, so we don't need it anymore. At least, until we see
issues pop up with older browsers.