ADD API user controller tests
Some basic user controller tests, only on non-signed-in and admin users.

Found out that you can view logging output by running 'pytest -s'. No
need to hack the log levels now, every level is displayed.

The user controller requires all requests to be from signed in users.
We're not doing a request to the sign in API as that should be tested
separately by the sign in controller. What we're doing instead is
bypassing the API request and directly generating the token that the
users get as part of sign in. This token is added to the header in
authed requests so the API thinks we're signed in.

This generated token is not currently a fixture, but we might want to
change it into one later when there's multiple user roles.

Fixtures in the new fixtures/ directory needs to be imported by
conftest.py in order to be picked up automatically by pytest.
Unfortunately, this has the side effect of the python linter not
picking this up and labelling these imports as unused imports.

A raw connection to the database is required sometimes to verify that
certain operations actually did what they're supposed to, so I've made
that connection a session scope fixture. Unfortunately, using this same
connection for the database cleanup causes errors, so we're not using it
for the reset_database fixture.

A fixture was also added for creating a temporary user that only lasts
for the execution of that 1 test case.