FIX UI nested css not working
Took out postcss-preset-env because I thought it was only needed for
purgecss. Turns out this isn't true, so I need to restore it.

Apparently, postcss is actually very bare bones, it doesn't support
nesting out of the box. postcss-preset-env was needed to add some basic
plugins to it, so we don't have to manually hunt down each plugin
ourselves. This includes the nesting plugin.