ADD UI can now add an user to a course
Implement a basic enrolment form that lets you select a user and a role
to add to a course.

Changed the API all for user enrolment to accept a single enrolment
instead of a list. I think initially, I wanted to be able to add a list
of users. But the basic implementation only adds 1 user at a time, so
I'm putting that off for later.

Added API tagging for the enrolment controller.

Added select elements to the formVertical css class.