EDIT UI set API base URL globally
While working on the new access system, I couldn't use the axios
instance to make requests to the API because it wouldn't pick up the
authorization header after login. Using the global axios import though
works fine. I suspect this has something to do with axios instances
picking up common settings at the time of instantiation, so because
we're using an axios instance that's created before sign in, we don't
have the authorization header. To make the workaround simple, I'm now
just using the global axios for requests.

Instead of only setting the baseURL on an axios instance, I now need to
configure it on the global axios. This means that the helper method for
generating API URLs used by the models is no longer needed and has been
removed.

One side effect of this change is that non-API requests now will need to
override the baseURL or they'll have /api prepended to their paths.