ADD API get the signed in user's courses & role
Doing the permission system on the UI side, need to know what courses the
signed in user is enroled in and what role they have in those courses.

On reflection, I'm not sure Token is the best place for the user context
to be organizationally. Since we're already retriving the user's courses
as part of the context, I've just reused the results for this new API
endpoint.

Also added tags and summaries documentation for the Enrolments
controller.