ADD UI simple role based permissions system
New vuex module for checking to see what roles a user has. It's not
really a permission system like we have in the API, since all it does is
tell you what role a user have. So instead of naming the module
'permission', I'm calling it 'access' for now. This should let us put in
a separate actual 'permission' system later if needed.

The access data is loaded on sign in and page load. It doesn't update if
roles are changed, we'll need to work that out later.

The isAdmin, isInstructor, isStudent getters can be accessed using
a component's computed property. An example can be seen in the Header
component in this commit. The isInstructor and isStudent getters require
a parameter to be passed in, so they need to be called as a function,
e.g.: this.$store.getters['access/isStudent'](courseId)